services/image-analyzer/requirements.txt
pika
uvicorn
easyocr
Pillow
python-dotenv
google-genai
jsonschema
torch==2.8.0
torchvision==0.23.0
